src/test/java/com/github/brunomndantas/jscrapper/support/parser/collection/text/reference/CollectionLongTextParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.collection.text.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class CollectionLongTextParserTest{

    @Test
    public void constructorTest() {
         CollectionLongTextParser parser = new  CollectionLongTextParser();
        assertEquals(null, parser.getDefaultValue());
    }

    @Test
    public void parseElement() throws Exception {
         CollectionLongTextParser parser = new  CollectionLongTextParser();

        assertEquals(2L, parser.parseElement(new DummyElement(){
            @Override
            public String getText() {
                return "2";
            }
        }));
    }

}
